+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 
-* Fixed where links wrapped in angle brackets (`<>`) weren't parsed.
+* Added rudimentary Discord embed parsing to the ChatGPT Plugin.
+* Fixed where the ChatGPT plugin didn't parse links wrapped in angle brackets (`<>`).
 * Added including channel name for ChatGPT prompts originating in public channels.
 * Added discord message character limit to ChatGPT instructions.
 * Added sending of typing signal while processing ChatGPT prompt.